--- a/pkg/virt-operator/kubevirt.go
+++ b/pkg/virt-operator/kubevirt.go
@@ -59,6 +59,7 @@ const (
 type KubeVirtController struct {
 	clientset            kubecli.KubevirtClient
 	queue                workqueue.RateLimitingInterface
+	delayedQueueAdder    func(key interface{}, queue workqueue.RateLimitingInterface)
 	kubeVirtInformer     cache.SharedIndexInformer
 	recorder             record.EventRecorder
 	stores               util.Stores
@@ -119,6 +120,9 @@ func NewKubeVirtController(
 		installStrategyMap: make(map[string]*install.Strategy),
 		operatorNamespace:  operatorNamespace,
 		statusUpdater:      status.NewKubeVirtStatusUpdater(clientset),
+		delayedQueueAdder: func(key interface{}, queue workqueue.RateLimitingInterface) {
+			queue.AddAfter(key, defaultAddDelay)
+		},
 	}
 
 	c.kubeVirtInformer.AddEventHandler(cache.ResourceEventHandlerFuncs{
@@ -445,7 +449,7 @@ func (c *KubeVirtController) genericAddHandler(obj interface{}, expecter *contro
 		if expecter != nil {
 			expecter.CreationObserved(controllerKey)
 		}
-		c.queue.AddAfter(controllerKey, defaultAddDelay)
+		c.delayedQueueAdder(controllerKey, c.queue)
 	}
 }
 
@@ -467,7 +471,7 @@ func (c *KubeVirtController) genericUpdateHandler(old, cur interface{}, expecter
 
 	key, err := c.getKubeVirtKey()
 	if key != "" && err == nil {
-		c.queue.AddAfter(key, defaultAddDelay)
+		c.delayedQueueAdder(key, c.queue)
 	}
 	return
 }
@@ -528,7 +532,7 @@ func (c *KubeVirtController) enqueueKubeVirt(obj interface{}) {
 	if err != nil {
 		logger.Object(kv).Reason(err).Error("Failed to extract key from KubeVirt.")
 	}
-	c.queue.AddAfter(key, defaultAddDelay)
+	c.delayedQueueAdder(key, c.queue)
 }
 
 func (c *KubeVirtController) Run(threadiness int, stopCh <-chan struct{}) {