--- a/pkg/virt-operator/kubevirt_test.go
+++ b/pkg/virt-operator/kubevirt_test.go
@@ -32,6 +32,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/util/intstr"
+	"k8s.io/client-go/util/workqueue"
 
 	"kubevirt.io/kubevirt/pkg/certificates/triple/cert"
 
@@ -234,6 +235,10 @@ func (k *KubeVirtTestData) BeforeTest() {
 	k.stores.ConfigMapCache = k.informers.ConfigMap.GetStore()
 
 	k.controller = NewKubeVirtController(k.virtClient, k.apiServiceClient, k.kvInformer, k.recorder, k.stores, k.informers, NAMESPACE)
+	k.controller.delayedQueueAdder = func(key interface{}, queue workqueue.RateLimitingInterface) {
+		// no delay to speed up tests
+		queue.Add(key)
+	}
 
 	// Wrap our workqueue to have a way to detect when we are done processing updates
 	k.mockQueue = testutils.NewMockWorkQueue(k.controller.queue)
@@ -1224,7 +1229,6 @@ func (k *KubeVirtTestData) makeApiAndControllerReady() {
 			if exists {
 				makeDeploymentReady(obj)
 			}
-			time.Sleep(time.Second)
 		}
 	}
 
@@ -1238,7 +1242,7 @@ func (k *KubeVirtTestData) makePodDisruptionBudgetsReady() {
 		for !exists {
 			_, exists, _ = k.stores.PodDisruptionBudgetCache.GetByKey(NAMESPACE + pdbname)
 			if !exists {
-				time.Sleep(time.Second)
+				time.Sleep(100 * time.Millisecond)
 			}
 		}
 	}
@@ -1257,7 +1261,6 @@ func (k *KubeVirtTestData) makeHandlerReady() {
 			handlerNew.Status.NumberReady = 1
 			k.daemonSetSource.Modify(handlerNew)
 		}
-		time.Sleep(time.Second)
 	}
 }