**What this PR does / why we need it**:

Improves the general execution speed of virt-operator from 167 seconds down to 26 seconds.
In addition the parallel ginkgo test execution is enabled which reduces the execution time to 10 seconds for the operator, and the virt-handler tests now need 50 seconds compared to more than 200 seconds.

For code coverage, parallel test execution of ginkgo is disabled to not confuse the coverage generator with the fork model which ginkgo uses for parallel runs.

**Which issue(s) this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close the issue(s) when PR gets merged)*:
Fixes #

**Special notes for your reviewer**:

**Release note**:

```release-note
NONE
```